Move homeboy to MEM1 and use linker scripts for external symbols (#3)
Currently for Wii, gz homeboy is loaded in MEM2 at 0x90000000.
For GC and Wii, I'd like to place it at the end of MEM1 instead
(just before 0x81800000, and the end to avoid shifting heap
allocations made by the emu e.g. to keep N64 RAM in place for
SRM setups). This is because there's no MEM2 for GC, and it makes
installing hooks easier for Wii anyway since powerpc `bl` is
limited to a +/- 32MB range and the rest of the emu code is in MEM1.
Currently, calling a homeboy function from emu code requires a longer
code sequence so we can't just replace a `bl`.

The new address is now 0x817F8000, which gives us 32KiB to work with
on Wii and 29.5KiB for GC (since for GC there's the file system table
at the end of RAM, at 0x817FF6A0 for MQ-J). Currently homeboy is 12KiB
so hopefully this is enough for the near future (and if it isn't we can
just move it up a bit).

This requires some gz build changes but I've tested that and it works.
I haven't tested kz, but kz already places homeboy in MEM1 so this change
should be fine there too.

Using linker scripts instead of hardcoding addresses is not technically
necessary for moving the code, but it makes vc.h easier to read and the
compiler generates nicer code.